+++ R/InsertTable.R
@@ -179,12 +179,12 @@ is.bigint <- function(x) {
 #' Insert a table on the server
 #'
 #' @description
-#' This function sends the data in a data frame or ffdf to a table on the server. Either a new table
+#' This function sends the data in a data frame to a table on the server. Either a new table
 #' is created, or the data is appended to an existing table.
 #'
 #' @param connection          The connection to the database server.
 #' @param tableName           The name of the table where the data should be inserted.
-#' @param data                The data frame or ffdf containing the data to be inserted.
+#' @param data                The data frame containing the data to be inserted.
 #' @param dropTableIfExists   Drop the table if the table already exists before writing?
 #' @param createTable         Create a new table? If false, will append to existing table.
 #' @param tempTable           Should the table created as a temp table?